Correct the matching script for DIFT
Script accepts all the flags required for the DIFT model. It also
doesn't requires the paths for the 2 images on which it will perform
matching. In case they were processed before and it already has it
features, it just reuses it.

For now, resizing is made only inside the DIFT itself. For convinience
it depict both images with the same width (the biggest between two) but keeping the aspect ratio.
You can also manage the NN process by setting ration_tresh value
(default: None) and use_mutal (default: False) for mutual NN. You can
use them separatly of together.